refactor: cleaning up

+++ app/Http/Controllers/Api/NotificationController.php
@@ -6,47 +6,49 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\NotificationResource;
 use Illuminate\Notifications\DatabaseNotification;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class NotificationController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
-     * 
-     * @param \Illuminate\Http\Request $request
+     *
      * @return \Illuminate\Http\Resources\Json\JsonResource
      */
     public function index(Request $request)
     {
         return NotificationResource::collection(
-            $request->user()->notifications()->take(15)->latest()->paginate()
+            $request->user()->notifications()->latest()->paginate()
         );
     }
 
     /**
      * Mark a given notification as read
-     * 
-     * @param \Illuminate\Notifications\DatabaseNotification
+     *
      * @return \Illuminate\Http\Response
      */
     public function markAsRead(DatabaseNotification $notification)
     {
+        $this->authorize('markAsRead', $notification);
+
         $notification->markAsRead();
 
-        return response()->noContent();
+        return response()->ok();
     }
 
     /**
      * Remove the specified resource from storage.
-     * 
-     * @param \Illuminate\Notifications\DatabaseNotification $notification
+     *
      * @param \Illuminate\Http\Response
      */
     public function destroy(DatabaseNotification $notification)
     {
-        // @TODO: Authorize this
+        $this->authorize('delete', $notification);
 
         $notification->delete();
 
-        return response()->noContent();
+        return response()->ok();
     }
 }
